+++ HISTORY.md
@@ -2,3 +2,5 @@ v0.1.2 (dev)
 ======
 + bcftools and gsort are now required.
 + output never goes to STDOUT
++ output contigs to header
++ add svtools to docker